+++ includes/resources.ejs
@@ -34,12 +34,16 @@
   </script>
   <script data-ad-client="ca-pub-7573886460664405" async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"></script>
 
-  <% if(page == 'stats'){ %>
+  <% if (page === "stats") { %>
     <link rel="preload" href="/resources/css/inventory.css?<%- fileHashes.css['inventory.css'] %>" as="style" onload="this.onload=null;this.rel='stylesheet'">
     <script async type="module" src="/resources/js/<%- fileNameMap['local-time'] %>"></script>
     <script defer type="module" src="/resources/js/<%- fileNameMap['stats-defer'] %>"></script>
     <script defer type="module" src="/resources/js/<%- fileNameMap['rich-item'] %>"></script>
     <link rel="preload" href="/resources/js/<%- fileNameMap['inventory-view'] %>" as="script">
+  <% } %>
+
+  <% if (process.env.NODE_ENV == "development") { %>
+    <script defer type="module" src="/resources/js/<%- fileNameMap['development-defer'] %>"></script>
   <% } %>
 
   <script defer type="module" src="/resources/js/<%- fileNameMap['common-defer'] %>"></script>
